third_party/WebKit/LayoutTests/media/effective-media-volume.html

<!DOCTYPE html>
<title>effective media volume</title>
<script src="../resources/testharness.js"></script>
<script src="../resources/testharnessreport.js"></script>
<div id="log"></div>
<script>
[0, 0.5, 1].forEach(function(volume)
{
    [false, true].forEach(function(muted)
    {
        test(function()
        {
            var audio = new Audio();
            audio.volume = volume;
            audio.muted = muted;
            assert_equals(internals.effectiveMediaVolume(audio),
                          volume * (muted ? 0 : 1));
        }, "volume: " + volume + ", muted: " + muted);
    });
});
</script>
